refactor(input): flatten macos module shim
macos.rs was a single `pub mod uckey;` line. The uckey module is
already cfg-gated to macOS at the declaration, so the extra directory
level added a path hop without isolating anything.
