--- crates/seance-input/src/macos.rs
@@ -1,1 +1,0 @@
-pub mod uckey;
